Refactor the AeadServiceImpl.
I suggest in the following we always follow this pattern when implementing Java RPCs. To implement an RPC called "performRpc", we implement an additional function with the same name, which returns a single Response, and throws a GeneralSecurityException if there is an error in testing.

Then, we always implement the function "performRpc" as follows:

  @Override
  public void performRpc(
      Request request, StreamObserver<Response> responseObserver) {
    try {
      Response response = performRpc(request);
      responseObserver.onNext(response);
      responseObserver.onCompleted();
    } catch (Exception e) {
      responseObserver.onError(e);
    }
  }

This is then always the same.

PiperOrigin-RevId: 469416340
Change-Id: I39e0b7f82a47ac2df743a42f35a55b29f0075636