Add a test_keys_container.
This can be used to add keys in a human readable way.

Contrary to what I planned, I didn't now check consistency of the template: I cannot do it completely, it shouldn't be a problem, and in fact a reviewer of a CL can also check it (though I don't think there's much harm if it's missing).

PiperOrigin-RevId: 474069846
Change-Id: I57f3992836f7e6cc119864174c902a332a4a39ba