Reject JWTs with duplicate map-keys in JSON objects in Python.
Such tokens are already rejected in C++ and Go, and we also want to reject them in Python (and later also in Java) to be consistent between the languages. So this change should be considered a bug fix.

This may potentially break some users that depend on this, but this is unlikely.

This is done by providing a "object_pairs_hook" to the JSON parser. This hook can now also be used for the string validation currently done by "validate_all_strings". It is preferable to use the hook since then we don't have to write a recursive function.

PiperOrigin-RevId: 475558732
Change-Id: Ib5d5caa2e5a15d757e942e249f59049b12830574