Only run cross-language KMS integration test with one envelope template.
Currently they are run with all available templates, which causes the test to time-out.

Doing these tests with all templates is not really needed, since we only want to check that the integration with the KMS works.

That envelope encryption works with all these templates is already tested using the fake KMS and doesnt need to be checked here again.

NOKEYCHECK

PiperOrigin-RevId: 497169277
Change-Id: Ideaafe17eef31c6d378453825bebf5776e434eb7