Add cross language tests specifically for AesGcmHkdfStreamingKeyTest.
We can branch these from the tests for AesCtrHmacStreamingKey since they are so similar. We only need to remove the HMac stuff since the way the tag is computed is fixed in AesGcm.

There are a few bugs: go/java allow non-standard hash functions, and go allows ciphertext_segment_size of size 2^31.

PiperOrigin-RevId: 511805861
Change-Id: I7c75861130f37c10e135419d5f3af82bc4e06d6a