Add warning to HC vault AEAD implementation and improve HC vault tests.
HC vault AEAD only works properly with keys that set "derived=true". So we
now use such a key in the cross-language tests.

But this AEAD fails for requests where associated_data is empty, which for example is always the case for envelope encryption. So we now test HC vault using two keys, one for non-empty and one for empty associated data.

PiperOrigin-RevId: 572483791
Change-Id: Ic9b2fa248e229ea5f8dbef552f84f990159a5192