Add kms aead tests with empty and with non-ascii inputs.
I had to add a special case for AWS, because it does not allow empty plaintexts.

And I now test invalid associated data in the same test with the valid associated data.
This way, we know that decrypt only fails because of the invalid associated data.

Also, parameterize tests more.

PiperOrigin-RevId: 588674534
Change-Id: I5baadf54714f06138b6451d1aa4bbb0d7d8f7c37