Only test some combinations in the "primitive_creation_test".
Previously, we would test all triples (lang, key_template, primitive). After this CL, we only test all pairs (lang, key_template) with the "correct" primitive, plus for each pair (lang, primitive1, primitive2) we test a template for primitive1 with primitive2.
The test is currently timing out rather often.

TBH I'm not sure the test is worth keeping: the main thing which isn't tested in any of the "creation_consistency_tests" is that we test keysets with wrong primitives, which doesn't feel like it's a very fundamental thing to test...

PiperOrigin-RevId: 596871066
Change-Id: I2788397bcba04d81df7acfd13c9e4eb3bae8e463