Add cross-language tests with KMS AEAD with Tink prefix.
In https://github.com/tink-crypto/tink-go/issues/10 it
was mentioned that KMS Envelope Keys can and are used with a TINK prefix. And the same is probably true for KMS AEAD keys.

Add this test to make sure that all languages that currently support this are consistent. At the moment it does not work in Java.

PiperOrigin-RevId: 617204124
Change-Id: I5203daf1fdea2ac883f3269e498057f679ed0ce2