Don't use EngineFactory in RsaSsaPkcs1VerifyConscrypt.
We now use Conscrypt to create the RSAPublicKey object.

Also remove a test-case from the cross-language tests of an RSA public key with a small factor. This key was previously accepted and now gets rejected. It is fine to remove this test case because it is not a valid public key, but we also don't expect an implementation to reject such a key.

Also, fix some lint errors.

PiperOrigin-RevId: 653188872
Change-Id: I3b9600a2daa1ef05d784a984c375b2c8b94834cc